Features: Join condominium features Description
Made de join condominium command, command handler, validator and responce.

Made de join condominium endpoint.

Added two method to the repository interface and class, one for geting and entity base on a filter and another to verify and entity through a filter expresion.

Made de create and update repository assing a value to the createdAt or UpdatedAt properties.

Added the CondominiumUser to the sets of the context.

Change the setter of the createdAt properties of the entities.